==> code/01b-data-collection-lyrics.ipynb <==
# Part 1B: Data Collection Lyrics Genius

from lyricsgenius import Genius
from requests.exceptions import Timeout
import json
import pandas as pd

token = token
genius = Genius(token,skip_non_songs=True, excluded_terms=["(Remix)", "(Live)"], timeout=60,
                retries=2, verbose=False)

songs = pd.read_csv('../data/audio_features.csv')
songs.set_index('id',inplace=True)
songs.head(5)

data_lyrics = []
for index in songs.index:
    data = pd.DataFrame()
    track = songs.loc[index]['track']
    artist = songs.loc[index]['artist']
    res = genius.search_song(title=track, artist=artist)
    if res != None:
        lyrics = res.lyrics
        data = {'id': index, 'track': track, 'artist': artist, 'lyrics': lyrics}
        data_lyrics.append(data)

df_lyrics = pd.DataFrame(data_lyrics)
df_lyrics.head(5)

pd.DataFrame(data_lyrics)

df_lyrics.to_csv('../data/lyrics.csv')
