==> code/02-data-preprocessing.ipynb <==
# Part 2: Data Pre-Processing

Import Libraries and lyric data collected from Part 1B.

import pandas as pd
import re
from afinn import Afinn
from nrclex import NRCLex

Import audio feature data from Spotify scraping (Part 1A) and lyric data from Lyric Genius (Part 1B). 

lyrics = pd.read_csv('../data/lyrics.csv')
lyrics.drop_duplicates(subset=['id'],inplace=True)
audio = pd.read_csv('../data/audio_features.csv')
audio.drop_duplicates(subset=['id'],inplace=True)

audio.shape, lyrics.shape

df = lyrics.merge(audio[['id','genre']], on='id')

df.genre.value_counts()

## New Features
- Create features for word counts
- Create features based on lexicons

df['lyrics'] = [re.sub(r'^.*?Lyrics', '', str(lyric)) for lyric in df['lyrics']]
df['cleaned_lyrics'] = [str(lyric).replace('\n',' ') for lyric in df['lyrics']]
df['cleaned_lyrics'] = [re.sub("\[.*?\]","",lyric) for lyric in df['cleaned_lyrics']]

df['title_length'] = [len(title) for title in df['track']]
df['lines'] = [str(lyric).count('\n') for lyric in df['lyrics']]
df['sections'] = [str(lyric).count('\n\n') for lyric in df['lyrics']]
df['verse_count'] = [str(lyric).count('[Verse') for lyric in df['lyrics']]
df['chorus_count'] = [str(lyric).count('[Chorus') for lyric in df['lyrics']]
df['dash_count'] = [str(lyric).count('" -') for lyric in df['lyrics']]
df['words'] = [len(str(lyric).split()) for lyric in df['cleaned_lyrics']]
df['unique_words'] = [len(set(str(lyric).split())) for lyric in df['cleaned_lyrics']]

df.head(3)

df = df[df.dash_count<3]

df['emot_positive'] = [NRCLex(lyric).affect_frequencies.get('positive') for lyric in df['cleaned_lyrics']]
df['emot_negative'] = [NRCLex(lyric).affect_frequencies.get('negative') for lyric in df['cleaned_lyrics']]
df['emot_anger'] = [NRCLex(lyric).affect_frequencies.get('anger') for lyric in df['cleaned_lyrics']]
df['emot_trust'] = [NRCLex(lyric).affect_frequencies.get('trust') for lyric in df['cleaned_lyrics']]
df['emot_disgust'] = [NRCLex(lyric).affect_frequencies.get('disgust') for lyric in df['cleaned_lyrics']]
df['emot_fear'] = [NRCLex(lyric).affect_frequencies.get('fear') for lyric in df['cleaned_lyrics']]
df['emot_joy'] = [NRCLex(lyric).affect_frequencies.get('joy') for lyric in df['cleaned_lyrics']]
df['emot_surprise'] = [NRCLex(lyric).affect_frequencies.get('surprise') for lyric in df['cleaned_lyrics']]
df['emot_anticp'] = [NRCLex(lyric).affect_frequencies.get('anticip') for lyric in df['cleaned_lyrics']]

df.head(5)

df['track_lower'] = df['track'].str.lower()
#remove words within brackets and paranethes in song titles 
df['track_lower'] = [re.sub("\[.*?\]","",title) for title in df['track_lower']]
df['track_lower'] = [re.sub("\(.*?\)","",title) for title in df['track_lower']]

df['track_lower'] = [str(title).replace('- single version','') for title in df['track_lower']]
df['track_lower'] = [str(title).replace('- single edit','') for title in df['track_lower']]
df['track_lower'] = [str(title).replace('- remix','') for title in df['track_lower']]
df['track_lower'] = [str(title).replace('- remix','') for title in df['track_lower']]

df['lyrics_lower'] = df['cleaned_lyrics'].astype(str)
df['lyrics_lower'] = df['lyrics_lower'].str.lower()

df['lyrics_contain_title'] = [df.iloc[index]['track_lower'] in df.iloc[index]['lyrics_lower'] for index in df.index]

df.lyrics_contain_title.value_counts(normalize=True)

## Removing Bad Records
Some lyric data contains track lists or miscellaneous content instead of lyrics. 

Example of bad records:

df[df.id == '5QifrqmnjHzdOuBAheeeNU']['lyrics'].item()

df[df.id == '0FYUTxJZvvOVLopuhusqYC']['lyrics'].item()[0:1000]

Get a list of all records suspected to be faulty based on numerical values.

excl = df[(df['cleaned_lyrics'].str.contains('12.')) & (df['verse_count']==0)]
excl_ids = [i for i in excl['id']]

Exclude these ids from lyric data and audio feature data. Make sure dataframes have the same records in each dataset.

df_lyrics = df[df['id'].isin(excl_ids)==False]
df_lyrics.shape

#subste dataset for audio features
song_ids = [i for i in df_lyrics['id']]
df_audio = audio[audio['id'].isin(song_ids)]
df_audio.shape

df_lyrics.to_csv('../data/data_lyrics.csv', index=False)
df_audio.to_csv('../data/data_audio.csv',index=False)
